Rename to `/ohlc`
