+++ src/routes/token/index.ts
@@ -3,7 +3,7 @@ import balances from "./balances/index.js";
 import transfers from "./transfers/index.js";
 import holders from "./holders/index.js";
 import tokens from "./tokens/index.js";
-import prices from "./prices/index.js";
+import ohlc from "./ohlc/index.js";
 
 const router = new Hono();
 
@@ -11,6 +11,6 @@ router.route('/balances', balances);
 router.route('/transfers', transfers);
 router.route('/holders', holders);
 router.route('/tokens', tokens);
-router.route('/prices', prices);
+router.route('/ohlc', ohlc);
 
 export default router;
